refactor(sidebar-background-animations): Refactoring of atoms, molecules, and animations in the background, sidebar, and animations
