--- src/app/components/Title.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const Title = () => {
-    return (
-        <div className="absolute top-[-22%] left-[10%] inset-0  flex items-center h-310 w-310 rounded-full justify-center"></div>
-    );
-};
